+++ app/scripts/controllers/dashboard_controller.js
@@ -3,7 +3,7 @@ import App from 'scripts/application';
 import DashboardView from 'scripts/views/dashboard/dashboard';
 import DashboardArticles from 'scripts/collections/dashboard_articles';
 
-export default class DashboardController extends Marionette.Controller {
+export default class DashboardController extends Marionette.Object {
   index() {
     const articles = new DashboardArticles();