+++ app/scripts/controllers/landing_controller.js
@@ -2,7 +2,7 @@ import Marionette from 'backbone.marionette';
 import App from 'scripts/application';
 import ContentLandingView from 'scripts/views/content/landing';
 
-export default class LandingController extends Marionette.Controller {
+export default class LandingController extends Marionette.Object {
   index() {
     App.mainRegion.show(new ContentLandingView());
   }