+++ app/scripts/controllers/navigation_controller.js
@@ -3,7 +3,7 @@ import App from 'scripts/application';
 import Session from 'scripts/facades/session';
 import NavigationLayout from 'scripts/views/layouts/navigation_layout';
 
-export default class NavigationController extends Marionette.Controller {
+export default class NavigationController extends Marionette.Object {
   initialize() {
     App.navigationRegion.show(new NavigationLayout());
   }