+++ app/scripts/routers/articles.js
@@ -1,5 +1,6 @@
 import App from 'scripts/application';
 import BaseRouter from 'scripts/routers/base';
+import Controller from 'scripts/controllers/articles_controller';
 import routes from 'scripts/helpers/routes';
 import { props } from 'scripts/decorators';
 
@@ -9,7 +10,9 @@ import { props } from 'scripts/decorators';
   appRoutes: {
     'dashboard/articles/:page': 'index',
     'dashboard/articles/:page/article/:id': 'show'
-  }
+  },
+
+  controller: new Controller()
 })
 export default class ArticlesRouter extends BaseRouter {
   initialize() {