+++ app/scripts/routers/navigation.js
@@ -2,6 +2,7 @@ import Backbone from 'backbone';
 import App from 'scripts/application';
 import BaseRouter from 'scripts/routers/base';
 import Session from 'scripts/facades/session';
+import Controller from 'scripts/controllers/navigation_controller';
 import routes from 'scripts/helpers/routes';
 import { props } from 'scripts/decorators';
 
@@ -12,6 +13,7 @@ import { props } from 'scripts/decorators';
 })
 export default class NavigationRouter extends BaseRouter {
   initialize() {
+    this.controller = new Controller();
     this.listenTo(Session, 'create destroy', this.onSessionChange);
     this.listenTo(Backbone.history, 'route', this.onNavigationChange);
   }